src/test/java/com/emse/spring/automacorp/api/BuildingControllerTest.java
package com.emse.spring.automacorp.api;

import com.emse.spring.automacorp.dao.BuildingDao;
import com.emse.spring.automacorp.model.*;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.hamcrest.Matchers;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import java.util.List;
import java.util.Optional;

@WebMvcTest(BuildingController.class)
public class BuildingControllerTest {
    /**
     * All the controller tests were implemented before adding the authentication part of the backend.
     * The tests worked before the addition but the modifications made on a mock test don't work.
     * So in order to privilege security over the unit testing (knowing the real testing does work),
     *  I decided to keep the unit test as they were before (a mock test attempt can be found at the end of
     *  RoomControllerTest class).
     * To use the unit testing without the added security level, you have to remove/comment :
     *  - from build.gradle.kts the lines
     *      implementation("org.springframework.boot:spring-boot-starter-security")
     * 	    testImplementation("org.springframework.security:spring-security-test")
     * 	- from application.properties the lines
     * 	    spring.security.user.name=user
     *      spring.security.user.password=password
     *  - SpringSecurityConfig and SecurityController code and imports
     *  - from RoomControllerTest the lines
     *      import static org.mockito.BDDMockito.given;
     *      import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
     *  - perhaps some other imports of the removed implementation of spring boot security that I've forgotten
     */

    // Spring object to mock call to our app
    @Autowired
    private MockMvc mockMvc;

    // The serializer used by Spring to send and receive data to/from the REST controller
    @Autowired
    private ObjectMapper objectMapper;

    // We choose to mock the DAO used in the REST controller to limit the scope of our test
    @MockBean
    private BuildingDao buildingDao;

    BuildingEntity createBuildingEntity(Long id, String name){
        SensorEntity outsideTemperature = new SensorEntity(SensorType.TEMPERATURE, "Outside temperature");
        outsideTemperature.setId(1L);
        outsideTemperature.setValue(20.8);
        BuildingEntity buildingEntity = new BuildingEntity(name, outsideTemperature);
        buildingEntity.setId(id);
        return buildingEntity;
    }

    @Test
    void shouldFindAll() throws Exception {
        Mockito.when(buildingDao.findAll()).thenReturn(List.of(
                createBuildingEntity(1L, "Building 1"),
                createBuildingEntity(2L, "Building 2")
        ));

        mockMvc.perform(MockMvcRequestBuilders.get("/api/buildings").accept(MediaType.APPLICATION_JSON))
                // check the HTTP response
                .andExpect(MockMvcResultMatchers.status().isOk())
                // the content can be tested with Json path
                .andExpect(
                        MockMvcResultMatchers
                                .jsonPath("[*].name")
                                .value(Matchers.containsInAnyOrder("Building 1", "Building 2"))
                );
    }


    @Test
    void shouldReturnNullWhenFindByUnknownId() throws Exception {
        Mockito.when(buildingDao.findById(999L)).thenReturn(Optional.empty());

        mockMvc.perform(MockMvcRequestBuilders.get("/api/buildings/999").accept(MediaType.APPLICATION_JSON))
                // check the HTTP response
                .andExpect(MockMvcResultMatchers.status().isOk())
                // the content can be tested with Json path
                .andExpect(MockMvcResultMatchers.content().string(""));
    }

    @Test
    void shouldFindById() throws Exception {
        BuildingEntity buildingEntity = createBuildingEntity(1L, "Building 1");
        Mockito.when(buildingDao.findById(999L)).thenReturn(Optional.of(buildingEntity));

        mockMvc.perform(MockMvcRequestBuilders.get("/api/buildings/999").accept(MediaType.APPLICATION_JSON))
                // check the HTTP response
                .andExpect(MockMvcResultMatchers.status().isOk())
                // the content can be tested with Json path
                .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Building 1"));
    }

    @Test
    void shouldNotUpdateUnknownEntity() throws Exception {
        BuildingEntity buildingEntity = createBuildingEntity(1L, "Building 1");
        BuildingCommand expectedBuilding = new BuildingCommand(buildingEntity.getName(), buildingEntity.getOutsideTemperature(), buildingEntity.getRooms());
        String json = objectMapper.writeValueAsString(expectedBuilding);

        Mockito.when(buildingDao.findById(1L)).thenReturn(Optional.empty());

        mockMvc.perform(
                        MockMvcRequestBuilders
                                .put("/api/buildings/1")
                                .content(json)
                                .contentType(MediaType.APPLICATION_JSON_VALUE)
                )
                // check the HTTP response
                .andExpect(MockMvcResultMatchers.status().isBadRequest());
    }


    @Test
    void shouldUpdate() throws Exception {
        BuildingEntity buildingEntity = createBuildingEntity(1L, "Building 1");
        BuildingCommand expectedBuilding = new BuildingCommand(buildingEntity.getName(), buildingEntity.getOutsideTemperature(), buildingEntity.getRooms());
        String json = objectMapper.writeValueAsString(expectedBuilding);

        Mockito.when(buildingDao.findById(1L)).thenReturn(Optional.of(buildingEntity));

        mockMvc.perform(
                        MockMvcRequestBuilders
                                .put("/api/buildings/1")
                                .content(json)
                                .contentType(MediaType.APPLICATION_JSON_VALUE)
                )
                // check the HTTP response
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Building 1"))
                .andExpect(MockMvcResultMatchers.jsonPath("$.id").value("1"));
    }

    @Test
    void shouldCreate() throws Exception {
        BuildingEntity buildingEntity = createBuildingEntity(1L, "Building 1");
        BuildingCommand expectedBuilding = new BuildingCommand(buildingEntity.getName(), buildingEntity.getOutsideTemperature(), buildingEntity.getRooms());
        String json = objectMapper.writeValueAsString(expectedBuilding);

        Mockito.when(buildingDao.existsById(1L)).thenReturn(false);
        Mockito.when(buildingDao.save(Mockito.any(BuildingEntity.class))).thenReturn(buildingEntity);

        mockMvc.perform(
                        MockMvcRequestBuilders
                                .post("/api/buildings")
                                .content(json)
                                .contentType(MediaType.APPLICATION_JSON_VALUE)
                )
                // check the HTTP response
                .andExpect(MockMvcResultMatchers.status().isOk())
                .andExpect(MockMvcResultMatchers.jsonPath("$.name").value("Building 1"))
                .andExpect(MockMvcResultMatchers.jsonPath("$.id").value("1"));
    }

    @Test
    void shouldDelete() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.delete("/api/buildings/999"))
                .andExpect(MockMvcResultMatchers.status().isOk());
    }

}
